movies

+++ Advanced/JS Applications/Single page application/01. SoftForum_ÉÑßpaß¿/02.Movies/src/login.js
@@ -14,39 +14,29 @@ function startLogin(){
 function verifyData(){
     let email = form.querySelector("#email").value.trim();
     let password = form.querySelector("#password").value.trim();
-
-    if(email == ""){
-        alert("Fill the email field.");
-    }else{
-        if(password == ""){
-            alert("Fil the password field.");
-        }else{
-            form.querySelector("#email").value = "";
-            form.querySelector("#password").value = "";
-            return {email,password};
-        }
-    }
-   
+          
+    form.querySelector("#email").value = "";
+    form.querySelector("#password").value = "";
+    return {email,password};              
 }
 
 async function submit(event){
+    event.preventDefault();
+
     let data = verifyData();   
-    console.log(data);
-    try {
-        let endpoint = "/users/login";
-        let options = {
-            "method":"post",
-            "headers":{"Content-type":"application/json"},
-            "body": JSON.stringify(data)
-        }
-        let req = await request(endpoint,options);
-        console.log(req);    
+    let endpoint = "/users/login";
+    let options = {
+        "method":"post",
+        "headers":{"Content-type":"application/json"},
+        "body": JSON.stringify(data)
+    }
+
+    let req = await request(endpoint,options);
+    
+    if(req){
         localStorage.clear();
         localStorage.setItem("user",JSON.stringify(req));
-
-        window.location = "index.html";
-    } catch (error) {
-        alert("Invalid email or password.");   
+        window.location = "index.html"; 
     }
 }
